Work on live alien research
live aliens are researchable, and give bonus research depending on rank.
there is no string template work going on, i've simply added fields for
what research it actually adds/uses for the ufopaedia lookup, and a list
of free researches you can get.
there NEEDS to be string work going on here, just not sure the best way
to go about it.
there is also a bug with overlapping ufopaedia articles, when you close
the first, it resets the pallette.
need to implement a check to make sure there isn't another one in the
queue