add cases for alien base assaults
these are merely the calls to generate a battlescape game from an alien base
only accessable via the "new battle" menu at this stage, but if you COULD fly to a base on the geoscape,
you would be able to generate one that way too. but at this point the geoscape won't spawn them.