Geoscape Functions iterate through all the bases, if we find a psi lab, train any allocated soldiers, and set the psi training flag to true.
if the psi training flag is set, pass that information to the "MonthlyReportState"
so it can pop up psi training once you're done looking at numbers.
(and more importantly switch palettes properly)


in the MonthlyReportState, popup the psi training menu on closing if the flag is set
also fixes the miscoloured button on the monthly report screen.