Xcom Activity Points for Interceptions now that multiple interception is in, i can comfortably add points for shooting down UFOs, and penalties for losing an interceptor.
with this, i believe all the scoring functions are in place.
funding should change monthly, country's dispositions will be
displayed properly. i'm not 100% sure i'm using the correct formulae
for calculating overall happiness and monthly score.
currently i'm averaging the the scores for all applicable regions/countries
and using the inverse of the formula for unhappiness to represent happiness.

the formula for pact signing is correct, as far as i can tell, but ufopedia
doesn't list the conditions for anything other than pact signing, so i'm sure
the rest is wrong (even though it DOES provide meaningful results)