small dogfight fixes double check to make sure we're literally on top of the UFO before initiating a dogfight. probably needs a tiny bit more leeway than being checked as a boolean expression, but solves the problem of dogfights starting 500 miles away from the target.
also, small change to the handling of music and battlemusic flags so they
get set properly.

wow that's a big description for having moved three lines of code
and adding a condition to an if statement on a fourth.

also, sorry for these rapid-fire commits, but small bugs like this keep
creeping in while i'm testing other things.