Monthly Report Bugfix 2
turns out replacing the integer variable with a function call was a bad idea.
so i'll move the function call to an integer variable and use that instead.
wait, i'm using the same function call and it works?
i can't add an integer that is the result of a function to an integer
that is the result of a function, unless i declare it as an integer first???

silly C++, go home, you're drunk.