Add Vertical Scale to graphs
oh god, this is a mess.
this class is starting to suffer some serious bloat issues,
but for the life of me, i can't figure out a better optimized method that
results in the text actually being where i want it.

if you can think of a better method... don't hesitate to send a PR.