better graph scaling
now we have nice round numbers on the graphs screens
using these methods to get the results i wanted makes me sick to the stomach
if someone can come up with a formula that does the same thing PLEASE send a PR,
because this feels like the worst optimized piece of shitty code i've submitted.
i could substitute the series of else-ifs to a case statement, but that's just
a different way of dealing with the problem, not a solution.