Make buttons prettyful.
now that we have too many buttons, let's make them all do something.
for starters, let's make the graph selection buttons display the right set
of buttons, and let's create some variables/pointers to use further on down
the road (for example, an array of which buttons are currently on)

then, let's make the country/region buttons toggle their colours
on and off when you click on them.