Line Drawing Method
Now that we have prettyful buttons that toggle boolean flags, let's draw
some dots to try and figure out how to plot the information accurately
on the graph.

while it may not look like much, this represents the result of WAY too
many hours and approximately 8 cups of coffees worth of testing with
variables and adjusting the formula and then retesting and readjusting
and yelling and geting frustrated and hitting things until they behaved.

protip: if you're looking at a country graph, and you can't get the data
to line up... double check that you're using COUNTRY data and not REGION
data, because i can't even tell you how long it took me to notice.