Line Drawing Method
now that we have the pixels lining up nicely and where we want them,
and the buttons are all doing things, let's try drawing some lines, and use
a bunch of surfaces that we can toggle with the boolean flags that the
buttons are setting for us.

what the hell. that worked and it took like 5 minutes.
do i draw my powers from exhaustion or something? because i'm damned if i
can code like this with a clear head.

anyway, that's one working graph screen, it should basically be copy/pasta
from here on out, then i can wrap a bow on this bitch and call it a day.