Added trajectory data to the ruleset.
Thanks to Volutar we now have the trajectory data in our ruleset. We are not
using it yet, but it's there.

Important: The original data had cities as part of zone 3, I have changed it to
zone 0, since we want to keep our existing rules intact. So every region with
cities has them as zone 0 and the rest of the zones get a bump on their index.
Regions without cities (eg Pacific) have their zone 3 areas moved to zone 0, and
the zones renumbered.