Landmark: everything but finances.
comitting at this point because everything is more or less working how it should
with the exception of the finances window, which uses a completely different
set of buttons and will likely require some refactoring and save game breaking
because we'll have to store a LOT more things as vectors rather than integers
notably: facility, craft and personnel maintainence. or something. i have to
check the original so i'd know for sure what needs added, but at this point i'm happy.