Graphs finished
added another save field "maintenance" to keep track of monthly maintenance
costs so we can display them on the graph.

at this point everything is WORKING and ready to send upstream, however i
feel like there's a bit of optimization in order here, as i'm making a lot
of redundant calls and setting variables way more often than needed.

the horrible mess that is the code at this point represents my approach to
coding. "whatever it takes" seems to describe it nicely.