Add Cydonia button if conditions are met.
add the Cydonia destination button if the appropriate conditions are met.
add the confirmation screen when choosing cydonia as a destination.
tweak the briefings a little to make room for the new mission types.
make the victory conditions exclusive to the actual game and not "new battle"
rename the final mission so we get the correct briefing between missions.
etc. etc. etc.

I think this might be that mystical 0.5.0