Proper Melee for large units
make the isBlocked function of pathfinding public.
start using it in melee calculations to prevent attacking through walls and windows.
make sure the target is VISIBLE to the attacker, too, to prevent attacking thru doors.

calculate a better target for pathfinding for melee units to match these extra conditions.
make allowances in this pathfinding for large units.

make allowances for large melee units in the action menu (mind controlled reapers)