Better handling for melee units
BattleUnit:
add a "charge target" (no need to save, as it only exists for the movement cycle)
and some functions to read/write it.

AggroBAIState:
melee units should only pick a destination they can reach AND attack from.
if they can't find a suitable target, they should fall back to "hide"
behaviour.

UnitWalkBState:
if we are charging a unit, run headlong towards them, ignoring any new targets
we may spot along the way (our target might drop in and out of sight along the way)
and face towards the target at the end of the route, in the hopes of landing an attack.
melee units can move and attack in the same action, allowing them to charge and
attack up to three times per turn. works for me.

Pathfinding:
fix the "findreachable" method to only return tiles that are actually reachable.
i can find no reference to this method anywhere else in the code, so this shouldn't
cause any weird behaviour anywhere else.

TileEngine and ProjectileFlyBState:
fix typo in "validMeleeRange"