start manipulating the variable
oooh, yeah, just like that.
a little to the left.

an alien spots a soldier, set him "visible" to the aliens for a number of
turns relative to the alien's intelligence.

decrement the counter at the beginning of each XCom turn, and set it to 0
if the unit dies.

if a soldier panics or goes berserk, give away his position.