psi aliens and launchers
they should revert to blaster launcher target if applicable, as there won't
be a chance for them to pick another target for it if they decide not to use
Psi. also, they pick their targets from the list of "exposed" soldiers,
not from team LOS, when calculating targets for such attacks.