fix research functions to account for live aliens correctly.
aliens are defined as anything that unlocks alien origins.
leaders are defined as anything that unlocks "leaderPlus"
commanders are anything that qualifies as a live alien that also unlocks the cydonia dependancy

i hate to hard-code these checks, because they aren't tftd-compatible, and it adds
a whole bunch of conditions on how to define things properly, modders will need a handbook.
:(