New option in options.cfg: canTransferCraftsInAirborne Default: OFF
Have you ever dreamed that you can transfer a craft while it is airborne? I mean when there is a base - with a free hangar - it can reach with fuel, which is closer then the original base. And this way the craft does not have to reserve much fuel, so it can use the remaining fuel for intercepting.

Now you can do this. :)
Just turn on canTransferCraftsInAirborne in options.cfg, and you can transfer that craft to ANY base it can reach with its remaining fuel.
Of course if a craft can't reach the selected destination base, it won't show up in the transfer-items-screen.
And of course transfer can only be done if the destination base have a free hangar. (not momentarily-free, completely-free!)
These type of transfers does not cost money, since the craft goes using its own fuel. :)
Crafts currently on the base are not affected, the normal rules applies to them.