fix for elevator blackout.
after looking into the map data, it seems that microprose were good enough
to make the grav lifts in the starting areas of alien base missions unique
in that they were the only grav lifts to contain the shiny stripy bits on
the second floor. this made them unique enough to flag as "discovered" in
the mission generator.