yet another fix for research popups.
if the base's available research has dependencies on this research,
or if the base's available research is unlocked by this research,
and it isn't unlocked by another discovered research or it's dependencies,

then unlock this research, and possibly it's dependencies, depending on if
the discovered research unlocks the dependent research in question.

simple, right?