Aliens now pick up weapons.
didn't need a whole AI state for this, it turns out.
it's just a handful of routines, so i stuck em in BattlescapeGame
so that HandleAI() could make use of them.

of course, it still doesn't ever HAPPEN because the data values are too low,
but that's nothing to do with me.