code clean and standardization
following an enquiry regarding difficulty and it's effects, something struck
me about how we were handling the data in various places.
sometimes we used a static_cast<int> and sometimes we cast (int) and
sometimes we did neither.
i figured some kind of standard was in order, so i settled on the middle ground
and cast them all as (int)