fix soldiers not spawning at alien bases.
refactored the code to find a valid spawn point to it's own function and
called it when setting up xcom units in alien bases, and when placing civilians

fixes the following:

units not showing up when taking 26 soldiers to an alien base in an avenger.
too few civilians showing up at terror sites.
0 civilians should also exist as a possibility.
initial FOV wasn't being calculated in base defense/assault missions.