implement ruleset-based MCD patches
while this class does require expansion to accommodate everything that
defines a mapData object, it does provide a "good start" that should be easy
to copypaste all the rest later on. for now i'm just adding in some bigWall
integers to differenciate them a bit, for the improvements to the
pathfinder.