slightly massive changes to sprite drawing routines.
adjusts dual wielding routine to properly represent dual wielding two of the
same weapon.

if two dual-handed weapons are held, choose the one currently being used
to display.

rifles will now be held properly, with the right arm "in front" of the gun,
as a human being would hold a rifle.

if a pistol and a 2 handed weapon are used, represent both when firing.
- this includes holding a rifle the left hand.