Fixes rare crash when deleting facility in the middle of base
Deleting a base facility at x=2 y=3, so roughly where the ok button is
in the confirm deletion box, can occasionally cause a viewMouseOver
event to fire where the facility just was.  As the facility no longer
exists its coordinates are no longer reliable.  Checking the coordinates
are sane avoids a potential crash.