Modability features.
assign each extra ruleset it's own internal ID (increments of 1000)
offset all the sprite and sound references defined by the ruleset by this
amount, in order to avoid mod conflicts.

this also adds a "list weight" to items, crafts, and ufopedia entries
and so on (these work in increments of 100) and by extension,
sorts modded items to appear where ever you want in any list.