_smoke acts as _intensity when _fire > 0
rewrites a bunch of smoke and fire functions. smoke may differ slightly
from the original, as there was a bug to do with the "averaging" the smoke
values at the end of the turn, where _smoke values were clamped to 15,
and then divided by _overlaps.
(clamp should only apply if _overlaps <= 1, and be applied after the
division in any other scenario)
smoke and fire can no longer occupy the same tile, but fire will obscure
line of sight.