Fixes bug 208
UnitDieBState is expecting units to be turning, collapsing, or standing.
If a unit is facing direction 3 then it does not turn for a death
animation so the unit keeps its current state; if this state is a non
standing state then UnitDieBState::think() cannot end so the game hangs.

Fix it so that all non collapsing and non turning units are treated as
dying.