New attribute in the manufacture section of the ruleset: producedItems
For the sake of moddability.

It is a map with name-value pairs. (just like the requiredItems)
If not specified, it contains only one element, which is the name-parameter with the value of 1.
The name specifies the produced object, and the value specifies how many of that object is created. (by completing "one object" of that project)

With this new attribute we can create manufacture-rules - for example - to convert 1 piece of Heavy Plasma Clip to 3 pieces of Elerium-115. (that clip is assembled of 3 pieces of elerium)
Sample ruleset file for this clips -> elerium conversion can be found here: http://openxcom.org/forum/index.php/topic,1623.msg15820.html#msg15820