smooth scrollin' bullet cam
not gonna put this in the options menus, since it's getting overhauled,
and that's just gonna create more work. not gonna add strings for it either
because i'm not sure what the implementation will be.

for the time being, to see this feature in action you will need to add the
following line to your options file:

battleSmoothCamera: true