fix stuck sell screen with storage limits
storage limits includes items on craft and in transfer towards total storage,
and compares it to facility storage. this discrepency causes the user to
become "trapped" in the sell screen after a particularly bad base defense in
which he loses storage facilities (rare, but possible.)

solution in this situation is to allow sale of items on-board the craft and
in transfer to free up the needed space.