Two bugs in dogfight about chance to hit
1) Chance to hit a UFO should be adjusted according to UFO size, the
formula is CTH = (1+3/Size)/2*Accuracy, where Size is 5 for Very Small
to 1 for Very Large. This is how it was in original game, in OXC it is
just flat right now.

2) Projectiles try to impact again and again because missing does not
prevent a check for impact. Fixing this required changing the movement
code, because previously when projectile reached max range it would
immediately be considered to have missed, which would prevent
projectiles hitting at max range (for example, when cautious or standart
attack is chosen)