more melee for aliens
units may now have a built in melee weapon that doesn't occupy a slot,
as defined in the unit section of the ruleset under the tag "meleeWeapon:"
with an accompanying string referring to an item.

this is only available to the AI, and as such some adjustments have taken
place to make aliens decide which to use, given their current situation,
as per original TFTD.

units can have a list of items that are "built in", and these items will
be given to the unit automatically. additionally, fixed weapons and alien
weapons can now be allocated to the left hand, meaning HWPs can have
secondary weapon systems installed.