functionality for mission objectives
allows for tile types _other than_ alien brains to count towards objectives
unifying alien base control modules, synomium devices and alien brain
handling into a common set of functions.

using "objectiveType" in the deployment section for a missionSite will
define which type of MCD to check for when tallying mission objectives.

using "objectivesRequired" allows definition of how many of said tile
types must be destroyed in order for the mission to be considered a success
note that this is a MAXIMUM value, and for practical reasons we will use
the lower between this number and the number of actual objects on the map,
assuming this is non-zero.

"objectivePopup" will define the message that will appear between turns,
after the objective requirements have been fulfilled.

"objectiveComplete" and "objectiveFailed" are both in the form [string, int]
and represent the string and score adjustment to be shown on the debriefing
screen when one or the other of the appropriate condition is met.

"despawnPenalty" represents how many points the aliens will receive when this
site despawns due to XCom negligence.

"points" defines how many points the aliens will receive each half hour that
the mission site stays active. this does not stack with the despawn penalty.

if someone wants to add this all to the wiki i'd be most grateful.