Fixed lag issue when research project is completed
Whenever the available research list is produced, the game halts. This
occurs when the new research list is opened as well as when a research
has been completed.

The issue comes from the fact that we check if the research project is
the same name as a unit. Since most research projects are not, an error
occurs as the unit could not be found. The generation of the error
causes the game to lag.

I've changed the code to check for live aliens only when needed as well
as checking if the unit exists instead of getting the unit and ensuring
an error was produced.