use dummies to pad out the array to solve direct reference issues
if we've culled nodes from the rmp, we'd have to retroactively adjust all
the built-in node links to reflect the nodes that were removed.
as this is an implementational nightmare, i've opted instead to pad out the
vector with dummy entries to keep all the sane links sane, and prune any
links to the dummy nodes. the dummy nodes will be ignored by the AI and
spawning functions completely.