Fix invalid _targetVoxel when LOF is blocked
Given the following situation:

	 012345678
	0   ____
	1  |    |
	2  |    |B
	3  |____|
	4      A

A takes an autoshot at B, which is visible, but LOF is blocked by a
wall to the north.

The first shot removes the northern wall at (6, 4):

	012345678
	1  ____
	2 |    |
	3 |    |B
	4 |___ |
	5     A

However, the second and third shots fly NW, perpendicular to the path to
B, removing the walls at (2, 2) and (3, 1)!

	012345678
	1   ___
	2      |
	3 |    |B
	4 |___ |
	5     A

In fact, the projectiles are flying towards (-16, -16, -24) due to
commit 62ec237f08530295a419e67a5a6b4336abc0df04.

This patch fixes this behavior by aiming at the middle of the target
tile in this case.