Fix a leak of BattleUnitKills if the debrief is not shown
The destructor for a BattleUnit doesn't delete the BattleUnitKills if
there is a valid Soldier* for the unit, on the assumption that the
object will pass ownership to the SoldierDiary, but that ownership move
only occurs on the DebriefingState, so if you happen to leave a battle
in a way that doesn't show that it will leak (e.g. 'Abandon Game')

This changes the BattleUnit destructor to unconditionally destroy any
remaining BattleUnitKills objects, and the SoliderDiary::updateDiary()
function to empty the BattleUnit's list when it takes ownership.