take two of previous commit
retain save/load/exit functionality but disable access to options when the
battlescape state queue is populated.

saving options creates a new battlescape state, overwriting the old one,
and the old state queue disappears into the aether, causing problems with
actions not being able to fully resolve, leaving units "stuck" and
potentially freezing the game.

however, should the game freeze mid-action for whatever other reason, it is
preferable to allow the player to save/load or exit the game, regardless of
the state queue.