Implemented simple hack-fix for intercept wrong way bug.
The hack/simple fix keeps the interceptor from looking ahead more than
halfway across the globe. The wrong way interception bug problem is caused
by that calculateMeetPoint may otherwise find an intercept point more than
halfway across the globe, and then calculateSpeed takes the shortest path
to it, which is the wrong way.

However, due to its simplicity, this fix will have trouble if an interceptor
is set to intercept a UFO that is on the far side of the world and is moving
slowly enough that the intercept point is also on the far side. It will take
the long way around in that case instead of meeting the UFO head on. It also
won't optimally intercept a faster UFO heading towards the interceptor at an
angle.