fix potential bug in dogfights
previous commit could have played weird when long range weapons
are disabled and engagement distance is set to cautious.

moved check for weapon disabled state to compensate, and set
projectileInFlight flag more appropriately.
this will be applied whether the weapon fired is a beam or missile,
as beam weapons have a burn duration, and only do the hit calculation on
the last frame.