Fix handling of unconscious ignored units
Let's assume a 2-stage alien colony mission.
When an unconscious unit is left behind on stage 1, while stage 1 is aborted, and all xcom units in stage 2 are killed... the game not only lets the craft survive, but also gives you points for destroying the control center and effectively removes the alien base from geoscape.

This is now fixed and mission is considered lost.