Stop the charade
The windows build is using MSVC and under macOS the `gcc` binary is a
symlink to clang, so effectively we were compiling it with the same
compiler twice.
The TravisCI build suffers from the same problem, btw.

We're also not using the matrix expansions, as we list all of our
potential builds under include explicitly. So just use that going
forward. To not duplicate the list of required packages, the apt
packages have been moved into a single list during install phase.

Switch to only uploading the bin/ directory, now that debugging build
breakages has finished. Do something special for macOS.