Apply clang-tidy fixes for performance-type-promotion-in-math-fn
floor(3) as used previously is the libc floor, which takes a double
always, so the cast to float is superfluous. floorf(3) would take a
float argument, but instead of switching to that, use std::floor which
has the proper overload for double/float, and as of C++11 also has one
for integral types, so use that directly.

The clang-tidy fixes were massaged manually, as it wasn't quite as
smart.