calculateLine:FastVoxelTraversal+GridIntersection
1. Fast Voxel Traversal Algorithm implemented for calculateLine

1.1 This increases accuracy of tile traversal, it will process
all tiles that the line passes through, one exception might be
if the line is completely on a point of a grid, then it will choose
one or the other.

1.2 This can also increase speed.

2. Line/Map/Grid/Box intersection implemented.

2.1 In case lines or points go outside the map/grid than they
are clipped towards the map/grid.

3. Other rare cases are single points, points inside or outside
the grid, one or both line points on a grid boundary, these are all
handled by this new code.

4. An attempt was made to move code for "process voxel" into a seperate
routine but this failed. Perhaps somebody else could try this.