Add Geoscape option to rearm/refuel before repairing
Before this change, the "Force craft launch" option is highly
restricted by repair time. Default behavior repairs, then rearms,
then refuels the craft. In case of a seriously injured Avenger,
repair time may be more than a month, and the craft can't be refueled
during that time. So forced launches for that period work only until
it uses up the fuel tank. At most a single tank for a month.